feat: regexp,Joi validation, error middleware

+++ controllers/cards.js
@@ -1,57 +1,70 @@
 const Card = require('../models/card');
-const sendError = require('../helpers/sendError');
-const checkIdValidation = require('../helpers/checkIdValidation');
+const ForbiddenError = require('../errors/forbidden-err');
+const NotFoundError = require('../errors/not-found-err');
 
-const readCards = (req, res) => {
+const readCards = (req, res, next) => {
   Card.find({})
     .populate(['likes', 'owner'])
     .then((cards) => res.send(cards))
-    .catch((error) => sendError({ error, errorText: 'Ошибка на стороне сервера', res }));
+    .catch((error) => next(error));
 };
 
-const createCard = async (req, res) => {
+const createCard = async (req, res, next) => {
   const { name, link } = req.body;
   try {
     const newCard = await Card.create({ name, link, owner: req.user._id });
     newCard.populate('owner').execPopulate()
       .then((card) => res.send(card));
   } catch (error) {
-    sendError({ error, errorText: 'Переданы некорректные данные при создании карточки', res });
+    next(error);
   }
 };
 
-const removeCard = (req, res) => {
-  checkIdValidation({ res, id: req.params.cardId, errorText: 'Переданы невалидные данные id карточки' });
-  Card.findByIdAndRemove(req.params.cardId)
-    .orFail()
-    .then(() => res.send({ message: 'Пост удалён' }))
-    .catch((error) => sendError({ error, res, errorNotFoundText: 'Запрашиваемая карточка не найдена' }));
+const removeCard = (req, res, next) => {
+  Card.findById(req.params.cardId)
+    .then((card) => {
+      if (!card) {
+        throw new NotFoundError('Запрашиваемая карточка не найдена');
+      }
+      if (card.owner._id.toHexString() !== req.user._id) {
+        throw new ForbiddenError('Недостаточно прав');
+      }
+      return Card.findByIdAndRemove(req.params.cardId)
+        .then(() => res.send({ message: 'Пост удалён' }));
+    })
+    .catch((error) => next(error));
 };
 
-const likeCard = (req, res) => {
-  checkIdValidation({ res, id: req.params.cardId, errorText: 'Переданы невалидные данные id карточки' });
+const likeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $addToSet: { likes: req.user._id } },
     { new: true },
   )
-    .orFail()
     .populate(['likes', 'owner'])
-    .then((card) => res.send(card))
-    .catch((error) => sendError({ error, res, errorNotFoundText: 'Запрашиваемая карточка не найдена' }));
+    .then((card) => {
+      if (!card) {
+        throw new NotFoundError('Запрашиваемая карточка не найдена');
+      }
+      res.send(card);
+    })
+    .catch((error) => next(error));
 };
 
-const dislikeCard = (req, res) => {
-  checkIdValidation({ res, id: req.params.cardId, errorText: 'Переданы невалидные данные id карточки' });
+const dislikeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $pull: { likes: req.user._id } },
     { new: true },
   )
-    .orFail()
     .populate(['likes', 'owner'])
-    .then((card) => res.send(card))
-    .catch((error) => sendError({ error, res, errorNotFoundText: 'Запрашиваемая карточка не найдена' }));
+    .then((card) => {
+      if (!card) {
+        throw new NotFoundError('Запрашиваемая карточка не найдена');
+      }
+      res.send(card);
+    })
+    .catch((error) => next(error));
 };
 
 module.exports = {
